Replace HTML in messages with plain text
Two messages included complex HTML. This is generally discouraged
(see
https://www.mediawiki.org/wiki/Help:System_message#Avoid_markup_that_doesn't_need_to_be_translated
). In addition, the automatic tests don't allow automatically merging
translations with <span> elements that also have the "style" attribute
for security reasons, so each translation of these messages
requires manual review.

Using <abbr> is allowed by the automatic tests, and it would
perhaps this scenario, but it is also not so recommended for
accessibility reasons (see
https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/abbr
)

So it's easier to simply spell out the abbreviation in parentheses
instead of using any markup.

Change-Id: Ic8dbc167a04807422c85602e6eced8d2f4e379a8